requirements.txt
Flask==2.3.6
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.6
Werkzeug==2.3.8
pandas==2.1.1
scikit-learn==1.3.2
numpy==1.27.5
python-dotenv==1.0.1
openai==1.7.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.6
Werkzeug==2.3.8
pandas==2.1.1
scikit-learn==1.3.2
numpy==1.27.5
python-dotenv==1.0.1
openai==1.7.0
